Revert "[eventfd] Make read()/write() work from ISRs, in theory."
This reverts commit 4377618eb498aa2ed3b7a71e74c18d1b136187d8.

That's not going to work, since newlib _lock_t's are used, and those are
FreeRTOS mutexes (semaphores).

We can replace our lock with a critical section, but the FD table is
still protected by a _lock_t. We'll just have to offer an alternate API
for write() (and maybe read()).